WIP: FeatureModel abstraction layer for feature coloring
