+++ src/main/java/org/mastodon/revised/ui/coloring/feature/FeatureProjectionId.java
@@ -1,4 +1,4 @@
-package org.mastodon.revised.ui.coloring.feature.revised;
+package org.mastodon.revised.ui.coloring.feature;
 
 import org.mastodon.feature.Multiplicity;
