+++ src/main/java/org/mastodon/revised/ui/coloring/feature/Projections.java
@@ -1,0 +1,10 @@
+package org.mastodon.revised.ui.coloring.feature;
+
+import org.mastodon.feature.FeatureProjection;
+
+public interface Projections
+{
+	public FeatureProjection< ? > getFeatureProjection( final FeatureProjectionId id );
+
+	public < T > FeatureProjection< T > getFeatureProjection( final FeatureProjectionId id, Class< T > target );
+}
