+++ src/main/java/org/mastodon/revised/ui/coloring/feature/TargetType.java
@@ -1,4 +1,4 @@
-package org.mastodon.revised.ui.coloring.feature.revised;
+package org.mastodon.revised.ui.coloring.feature;
 
 // TODO: Does equivalent of this enum already exist?
 public enum TargetType
